moved RHICommandlist code to cpp
